+++ include/utils/converter.h
@@ -3,13 +3,13 @@
 
 namespace Utils {
 
-using coordinate = std::pair<int, int>;
+using coordinate = std::pair<size_t, size_t>;
 
-inline int Convert2DIndexTo1DIndex(int row, int column, int row_count) {
+inline size_t Convert2DIndexTo1DIndex(size_t row, size_t column, size_t row_count) {
   return row*row_count + column;
 }
 
-inline coordinate Convert1DIndexTo2DIndex(int index1d, int rows, int columns) {
+inline coordinate Convert1DIndexTo2DIndex(size_t index1d, size_t rows, size_t columns) {
   return std::make_pair(index1d/rows, index1d%columns);
 }
 }
